Wheel bug fixes and improves
Fixed a bug where changing the rotation speed of the wheel and its spin time in the config did not work correctly (it showed that a different privilege was received than what actually occurred).
Removed the function for adjusting the wheel's rotation speed due to its lack of purpose. Instead, a function for setting a minimum number of spins has been added.
Added a Shuffle function in the config that randomizes all privileges.
Added a setting to adjust the "friction" of the wheel.
Added nonlinear animations for fire and improved the wheel animation.
Slightly changed the mechanics of the RANDOM wheel. Now, the actual privilege that the player will receive is placed in a random position.